+++ app/scripts/views/fields/field-view-url.js
@@ -4,7 +4,11 @@ var FieldViewText = require('./field-view-text');
 
 var FieldViewUrl = FieldViewText.extend({
     renderValue: function(value) {
-        return value ? '<a href="' + _.escape(value) + '" target="_blank">' + _.escape(value) + '</a>' : '';
+        return value ? '<a href="' + _.escape(this.fixUrl(value)) + '" target="_blank">' + _.escape(value) + '</a>' : '';
+    },
+
+    fixUrl: function(url) {
+        return url.indexOf(':') < 0 ? 'http://' + url : url;
     }
 });