+++ app/scripts/comp/drag-drop-info.js
@@ -1,0 +1,7 @@
+'use strict';
+
+var DragDropInfo = {
+    dragObject: null
+};
+
+module.exports = DragDropInfo;