+++ app/scripts/models/group-model.js
@@ -12,7 +12,9 @@ var GroupModel = MenuItemModel.extend({
         entries: null,
         filterKey: 'group',
         editable: true,
-        top: false
+        top: false,
+        drag: true,
+        drop: true
     }),
 
     initialize: function() {
@@ -134,6 +136,28 @@ var GroupModel = MenuItemModel.extend({
         }
         this.parentGroup.removeGroup(this);
         this.trigger('delete');
+    },
+
+    moveHere: function(object) {
+        if (!object || object.id === this.id) {
+            return;
+        }
+        if (object instanceof GroupModel) {
+            if (this.group.groups.indexOf(object.group) >= 0) {
+                return;
+            }
+            this.file.db.move(object.group, this.group);
+            object.parentGroup.removeGroup(object);
+            object.trigger('delete');
+            this.addGroup(object);
+        } else if (object instanceof EntryModel) {
+            if (this.group.entries.indexOf(object.entry) >= 0) {
+                return;
+            }
+            this.file.db.move(object.entry, this.group);
+            object.group.removeEntry(object);
+            this.addEntry(object);
+        }
     }
 });
 
@@ -143,7 +167,7 @@ GroupModel.fromGroup = function(group, file, parentGroup) {
     if (parentGroup) {
         model.parentGroup = parentGroup;
     } else {
-        model.set({ top: true }, { silent: true });
+        model.set({ top: true, drag: false }, { silent: true });
     }
     return model;
 };