+++ release-notes.md
@@ -2,6 +2,8 @@ Release notes
 -------------
 ##### v0.2.0 (not released yet)
 Bugfixes and new features  
+`+` improved open page ux
+`+` keyfiles from Dropbox  
 `-` #12: cannot edit entries without title  
 `+` #13: increase max generated password length