+++ release-notes.md
@@ -3,6 +3,7 @@ Release notes
 ##### v0.2.0 (not released yet)
 UX improvements, new features, bugfixes  
 `+` improved open page ux  
+`+` protected fields support  
 `+` keyfiles from Dropbox  
 `+` #17: option to hide entries from subgroups  
 `+` #5: groups and entries arranging